🚚 Move settings to models
