+++ lamindb/core/_settings.py
@@ -1,3 +1,3 @@
 # ruff: noqa: F401
 # backward compat
-from ..base._settings import Settings, settings
+from ..models._settings import Settings, settings
