+++ lamindb/core/datasets/_core.py
@@ -8,8 +8,8 @@ import anndata as ad
 import pandas as pd
 from upath import UPath
 
-from lamindb.base._settings import settings
 from lamindb.base.ids import base62
+from lamindb.models._settings import settings
 
 if TYPE_CHECKING:
     from mudata import MuData
