+++ lamindb/models/_feature_manager.py
@@ -37,7 +37,7 @@ from lamindb.models.record import (
 from lamindb.models.save import save
 from lamindb.models.schema import DICT_KEYS_TYPE, Schema
 
-from ..base._settings import settings
+from ..models._settings import settings
 from ._describe import (
     NAME_WIDTH,
     TYPE_WIDTH,
