+++ lamindb/models/save.py
@@ -13,7 +13,6 @@ from django.utils.functional import partition
 from lamin_utils import logger
 from lamindb_setup.core.upath import LocalPathClasses, UPath
 
-from ..base._settings import settings
 from ..core.storage.paths import (
     _cache_key_from_artifact_storage,
     attempt_accessing_path,
@@ -21,6 +20,7 @@ from ..core.storage.paths import (
     delete_storage_using_key,
     store_file_or_folder,
 )
+from ..models._settings import settings
 from .record import Record
 
 if TYPE_CHECKING:
