+++ tests/core/test_artifact.py
@@ -17,7 +17,6 @@ import pandas as pd
 import pytest
 import spatialdata as sd
 import yaml  # type: ignore
-from lamindb.base._settings import settings
 from lamindb.core.loaders import load_fcs, load_to_memory, load_tsv
 from lamindb.core.storage._zarr import identify_zarr_type, write_adata_zarr
 from lamindb.core.storage.paths import (
@@ -30,6 +29,7 @@ from lamindb.errors import (
     IntegrityError,
     InvalidArgument,
 )
+from lamindb.models._settings import settings
 from lamindb.models.artifact import (
     check_path_is_child_of_root,
     data_is_anndata,
